- Resolved unit test issues, curb that - Restructured folder structure
