+++ src/app/app.component.spec.ts
@@ -1,3 +1,4 @@
+import { ServiceWorkerModule } from '@angular/service-worker';
 import { MatListModule } from '@angular/material/list';
 import { TestBed, async } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
@@ -12,7 +13,10 @@ describe('AppComponent', () => {
         NoopAnimationsModule,
         MatListModule,
         MatSidenavModule,
-        RouterTestingModule
+        RouterTestingModule,
+        ServiceWorkerModule.register('', {
+          enabled: false,
+        })
       ],
       declarations: [
         AppComponent
